fix(study): delete failed when no dataset resource group was set (#444)
forgot to check if resource group name was set on study before attempting delete

+++ src/Sepes.Infrastructure/Service/DatasetCloudResourceService.cs
@@ -57,7 +57,11 @@ namespace Sepes.Infrastructure.Service
 
             try
             {
-                await _storageAccountService.DeleteStorageAccount(study.StudySpecificDatasetsResourceGroup, dataset.StorageAccountName, cancellationToken);
+                if (String.IsNullOrWhiteSpace(study.StudySpecificDatasetsResourceGroup) == false && String.IsNullOrWhiteSpace(dataset.StorageAccountName) == false)
+                {
+                    await _storageAccountService.DeleteStorageAccount(study.StudySpecificDatasetsResourceGroup, dataset.StorageAccountName, cancellationToken);
+                }
+              
             }
             catch (Exception ex)
             {
@@ -71,7 +75,10 @@ namespace Sepes.Infrastructure.Service
 
             try
             {
-                await _resourceGroupService.Delete(study.StudySpecificDatasetsResourceGroup, cancellationToken);
+                if (String.IsNullOrWhiteSpace(study.StudySpecificDatasetsResourceGroup) == false)
+                {
+                    await _resourceGroupService.Delete(study.StudySpecificDatasetsResourceGroup, cancellationToken);
+                }
             }
             catch (Exception ex)
             {
